requirements.txt
python-dotenv
asyncio

sqlalchemy 
alembic 
asyncpg

fastapi
uvicorn